refactor: update type definitions and improve error handling in form components
