+++ src/app/util/return-error-messages.ts
@@ -1,4 +1,5 @@
-export default function returnErrorMessages(errorObj: { errors: {} }): string[] {
-  return Array.from(Object.values(errorObj.errors));
+export default function returnErrorMessages(errorObj: { errors: Record<string, string[]> }): string[] {
+  return Object.values(errorObj.errors).flat();
 }
+
